feat: split config
