+++ package.json
@@ -23,22 +23,23 @@
     "prettier": "npx prettier --write \"**/*.{js,jsx,ts,tsx,css,scss,json,md}\""
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0-beta.15",
-    "@docusaurus/preset-classic": "^2.0.0-beta.15",
+    "@docusaurus/core": "^2.0.0-beta.18",
+    "@docusaurus/preset-classic": "^2.0.0-beta.18",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.0.0-beta.15",
+    "@docusaurus/module-type-aliases": "^2.0.0-beta.18",
     "@mdx-js/react": "^1.6.22",
     "@svgr/webpack": "^6.2.1",
-    "@tsconfig/docusaurus": "^1.0.4",
-    "docusaurus-plugin-sass": "^0.2.1",
+    "@tsconfig/docusaurus": "^1.0.5",
+    "docusaurus-plugin-sass": "^0.2.2",
     "file-loader": "^6.2.0",
-    "prettier": "^2.4.1",
-    "prism-react-renderer": "^1.2.1",
-    "sass": "^1.43.4",
-    "typescript": "^4.5.2",
+    "lodash": "^4.17.21",
+    "prettier": "^2.6.1",
+    "prism-react-renderer": "^1.3.1",
+    "sass": "^1.49.9",
+    "typescript": "^4.6.3",
     "url-loader": "^4.1.1"
   }
 }
